More clean up: renamed AtomicBool.hpp to Atomic.hpp.
